refactor(l2): split EutxoL2Screener out of EutxoL2Ledger
Screening is stateless (only config, no state/store), so give it its own class —
symmetric with the remote side, which already has a standalone RemoteL2Screener.

- New EutxoL2Screener(config) extends L2Screener[IO] with screenTx/screenDeposit.
- New EutxoDepositGates holds the two deposit-validity gates shared by the
  screener (screenDeposit) and the command path (registerDeposit).
- EutxoL2Ledger drops `extends L2Screener[IO]` and the two screen methods;
  keeps L2Ledger + EutxoL2LedgerReader (the reader is stateful).
- Serve's eutxo backend returns a fresh EutxoL2Screener in the screener slot;
  the head harness passes one into the MRM's screener slot.
